+++ src/main/java/roomescape/config/interceptor/AdminHandlerInterceptor.java
@@ -5,8 +5,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 import roomescape.entity.LoginMember;
-import roomescape.entity.Role;
-import roomescape.exception.ForbiddenException;
+import roomescape.global.Role;
+import roomescape.exception.custom.ForbiddenException;
 import roomescape.jwt.JwtExtractor;
 import roomescape.service.AuthService;
