+++ src/main/java/roomescape/domain/ThemeRanking.java
@@ -1,11 +1,12 @@
-package roomescape.entity;
+package roomescape.domain;
 
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import roomescape.entity.Reservation;
+import roomescape.entity.Theme;
 
-// TODO: repo JPA 적용 후, 제거
 public class ThemeRanking {
 
     private static final int RANKING_SIZE = 10;
