+++ src/main/java/roomescape/entity/Member.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import roomescape.global.Role;
 
 @Entity
 public class Member {
