+++ src/main/java/roomescape/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.global.dto;
+package roomescape.exception;
 
 public record ErrorResponse(
         String message
