+++ src/main/java/roomescape/exception/custom/InvalidMemberException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.custom;
 
 public class InvalidMemberException extends IllegalArgumentException {
