+++ src/main/java/roomescape/exception/custom/InvalidReservationException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.custom;
 
 public class InvalidReservationException extends IllegalArgumentException {
