+++ src/main/java/roomescape/exception/custom/InvalidReservationTimeException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.custom;
 
 public class InvalidReservationTimeException extends IllegalArgumentException {
