+++ src/main/java/roomescape/exception/custom/InvalidThemeException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.custom;
 
 public class InvalidThemeException extends IllegalArgumentException {
