+++ src/main/java/roomescape/exception/custom/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.custom;
 
 public class UnauthorizedException extends RuntimeException {
     public UnauthorizedException(String message) {
