+++ src/main/java/roomescape/service/ReservationService.java
@@ -11,10 +11,10 @@ import roomescape.entity.Member;
 import roomescape.entity.Reservation;
 import roomescape.entity.ReservationTime;
 import roomescape.entity.Theme;
-import roomescape.exception.InvalidMemberException;
-import roomescape.exception.InvalidReservationException;
-import roomescape.exception.InvalidReservationTimeException;
-import roomescape.exception.InvalidThemeException;
+import roomescape.exception.custom.InvalidMemberException;
+import roomescape.exception.custom.InvalidReservationException;
+import roomescape.exception.custom.InvalidReservationTimeException;
+import roomescape.exception.custom.InvalidThemeException;
 import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
